requirements.txt
Flask>=2.3,<3.1
Flask-SQLAlchemy>=3.0,<3.2
Flask-Login>=0.6,<0.7
Flask-WTF>=1.1,<1.3
Flask-Bcrypt>=1.0,<1.1
Flask-Migrate>=4.0,<4.1
grpcio>=1.50.0
grpcio-tools>=1.50.0
protobuf>=4.20.0 # Aligned with recent grpcio-tools
Werkzeug>=2.3,<3.1
python-dotenv>=1.0.0,<1.1.0
email-validator>=2.0.0,<2.2.0 # For WTForms email validation